Optimization: Token::tokAt
Do not check sign of index in each loop step and do not check this pointer.
This needs an additional correction of Tokenizer::simplifyMulAndParens which calls tokAt on null pointer.

This is only small improvement and maybe it is too risky to no longer allow tokAt on a null this pointer. Even though one should never call functions on a null this pointer.